@@ -1,15 +1,19 @@
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
+import { createRoot } from 'react-dom/client';
+import { StrictMode } from 'react';
 import { BrowserRouter } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import store from './store/index';
+import './index.css';
+import App from './App';
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(
   <BrowserRouter>
     <Provider store={store}>
-      <App />
+      <StrictMode>
+        <App tab='home' />
+      </StrictMode>
     </Provider>
-  </BrowserRouter>,
-  document.getElementById('root')
+  </BrowserRouter>
 );
